added bee2 package
Yeah, it's an item i made in beepkg v2.5 or whatever that online bee2 package maker is. pretty dumb stuff, i checked embed voxel and i didn't assign it an instance, so yeah there's either a leak in the map you use it in or it won't compile in the first place.